GPL-2.0 licenses: "Lesser|Library"
The LGPL license was originally called the "GNU LIbrary General Public License" and some license tests refer to the "original" term "Library" instead of "Lesser".

This change is to let license matchers yield a positive match for the usage of "Library" for the "GPL-2.0*" licenses. See also [JDK-8355388](https://bugs.openjdk.org/browse/JDK-8355388). This change does not add an `<alt>` to every occurence of "Lesser" in all licenses.